Refactor: anniversary 탭 훅 분리 #26
