+++ src/components/features/date-calendar/anniversary/anniversary-tab.tsx
@@ -1,21 +1,13 @@
 'use client';
 
-import { useState, useEffect, useCallback } from 'react';
-import { useAnniversaries } from '@/lib/hooks/use-anniversaries';
+import { useEffect } from 'react';
 import AnniversaryHeader from './anniversary-header';
 import AnniversaryList from './anniversary-list';
 import AnniversaryEditor from './anniversary-editor';
-import { Anniversary, PartnerInfo } from '@/types/anniversary.type';
 import { Dialog } from '@/components/ui/dialog';
-
-import {
-  DEFAULT_NICKNAME,
-  PLACEHOLDER_IMAGE,
-} from '@/constants/anniversary.constants';
-import {
-  getCoupleById,
-  getUserById,
-} from '@/lib/services/anniversary.services';
+import { useAnniversaries } from '@/lib/hooks/date-calendar/use-anniversaries';
+import { usePartnerInfo } from '@/lib/hooks/date-calendar/use-partner-infor';
+import { useAnniversaryActions } from '@/lib/hooks/date-calendar/use-anniversary-actions';
 
 type AnniversaryTabProps = {
   coupleId: string;
@@ -36,30 +28,17 @@ const AnniversaryTab = ({
     deleteAnniversary: remove,
   } = useAnniversaries(coupleId, startDate);
 
-  const [isEditorOpen, setIsEditorOpen] = useState(false);
-  const [editingAnniversary, setEditingAnniversary] =
-    useState<Anniversary | null>(null);
-  const [partner, setPartner] = useState<PartnerInfo | null>(null);
-
-  const fetchPartnerInfo = useCallback(async () => {
-    try {
-      const couple = await getCoupleById(coupleId);
-
-      const partnerId =
-        couple.userAId === userId ? couple.userBId : couple.userAId;
-
-      const partnerData = await getUserById(partnerId);
-
-      setPartner({
-        id: partnerId,
-        profileImage: partnerData.profileImage || PLACEHOLDER_IMAGE,
-        nickname: partnerData.nickname || DEFAULT_NICKNAME,
-      });
-    } catch (error) {
-      console.error('파트너 정보 가져오기 실패했습니다.:', error);
-      setPartner(null);
-    }
-  }, [coupleId, userId]);
+  const { partner, fetchPartnerInfo } = usePartnerInfo(coupleId, userId);
+  const {
+    editingAnniversary,
+    isEditorOpen,
+    setIsEditorOpen,
+    handleEdit,
+    handleDelete,
+    handleUpdate,
+    handleAddClick,
+    handleSubmitSuccess,
+  } = useAnniversaryActions(update, remove);
 
   useEffect(() => {
     const loadData = async () => {
@@ -72,41 +51,6 @@ const AnniversaryTab = ({
 
     loadData();
   }, [loadAnniversaries, fetchPartnerInfo]);
-
-  const handleAddClick = () => {
-    setEditingAnniversary(null);
-    setIsEditorOpen(true);
-  };
-
-  const handleEdit = (id: string) => {
-    const anniversaryToEdit = anniversaries.find((a) => a.id === id);
-    setEditingAnniversary(anniversaryToEdit || null);
-    setIsEditorOpen(true);
-  };
-
-  const handleDelete = async (id: string) => {
-    try {
-      await remove(id);
-    } catch (error) {
-      console.error('Error deleting anniversary:', error);
-      alert('기념일 삭제에 실패했습니다.');
-    }
-  };
-
-  const handleUpdate = async (data: Anniversary) => {
-    try {
-      await update(data);
-      handleSubmitSuccess();
-    } catch (error) {
-      console.error('Error updating anniversary:', error);
-      alert('기념일 수정에 실패했습니다.');
-    }
-  };
-
-  const handleSubmitSuccess = () => {
-    setIsEditorOpen(false);
-    setEditingAnniversary(null);
-  };
 
   return (
     <Dialog open={isEditorOpen} onOpenChange={setIsEditorOpen}>
@@ -123,7 +67,7 @@ const AnniversaryTab = ({
         <AnniversaryList
           anniversaries={anniversaries}
           onDelete={handleDelete}
-          onEdit={handleEdit}
+          onEdit={(id) => handleEdit(id, anniversaries)}
         />
 
         <AnniversaryEditor
